+++ packages/react/src/components/Box.tsx
@@ -3,8 +3,6 @@ import { styled } from '../styles'
 
 export const Box = styled('div', {
   padding: '$6',
-  borderRadius: '$md',
-  backgroundColor: '$gray800',
   border: '1px solid $gray600',
 })